Refactored to include a pl_sentiment package and a scripts directory.
