+++ pl_sentiment/__init__.py
@@ -1,0 +1,1 @@
+from .twitter_stream import TwitterStream
