+++ pl_sentiment/config/twitter_api_parameters.ini
