+++ scripts/stream_tweets.py
@@ -1,0 +1,8 @@
+"""
+Stream Tweets from the Twitter filtered stream API.
+"""
+from context import TwitterStream
+
+twitter_stream = TwitterStream()
+twitter_stream.reset_all_rules()
+twitter_stream.connect_to_stream()
